Convert time function to module
